+++ relay-redis/src/real.rs
@@ -1,5 +1,5 @@
+use deadpool::Runtime;
 use deadpool::managed::{BuildError, Manager, Metrics, Object, Pool, PoolError};
-use deadpool_redis::{ConfigError, Runtime};
 use redis::{Cmd, Pipeline, RedisFuture, Value};
 use std::time::Duration;
 use thiserror::Error;
@@ -34,10 +34,6 @@ pub enum RedisError {
     /// An error that occurs when creating a Redis connection pool.
     #[error("failed to create redis pool: {0}")]
     CreatePool(#[from] BuildError),
-
-    /// An error that occurs when configuring Redis.
-    #[error("failed to configure redis: {0}")]
-    ConfigError(#[from] ConfigError),
 }
 
 /// A collection of Redis clients used by Relay for different purposes.
@@ -104,7 +100,7 @@ impl AsyncRedisClient {
 
         // We use our custom cluster manager which performs recycling in a different way from the
         // default manager.
-        let manager = pool::CustomClusterManager::new(servers, false, opts.recycle_check_frequency)
+        let manager = pool::CustomClusterManager::new(servers, false, opts.clone())
             .map_err(RedisError::Redis)?;
 
         let pool = Self::build_pool(manager, opts)?;
@@ -122,8 +118,8 @@ impl AsyncRedisClient {
     pub fn single(server: &str, opts: &RedisConfigOptions) -> Result<Self, RedisError> {
         // We use our custom single manager which performs recycling in a different way from the
         // default manager.
-        let manager = pool::CustomSingleManager::new(server, opts.recycle_check_frequency)
-            .map_err(RedisError::Redis)?;
+        let manager =
+            pool::CustomSingleManager::new(server, opts.clone()).map_err(RedisError::Redis)?;
 
         let pool = Self::build_pool(manager, opts)?;